refactor(video-editor): extract audio-render and interactive-body units
The three touched files sit above the 800-line file-size ceiling (epic #4339) and may only shrink, so the added logic tripped the per-file ratchet. Extract focused units so each file ends below its frozen baseline:

- `audioTrackFrom*ForRender` + `resolveRenderAudioTracks` -> `video_editor_audio_render.dart` (shared by the provider and the render service)
- the interactive timeline body widget -> `video_editor_timeline_interactive_body.dart`

No behaviour change. provider 1051->1026, render_service 1065->1054, video_editor_timeline 1149->911.
