+++ mobile/lib/services/video_editor/video_editor_render_service.dart
@@ -13,6 +13,7 @@ import 'package:openvine/extensions/complete_parameters_extensions.dart';
 import 'package:openvine/models/divine_video_clip.dart';
 import 'package:openvine/services/crash_reporting_service.dart';
 import 'package:openvine/services/native_proofmode_service.dart';
+import 'package:openvine/services/video_editor/video_editor_audio_render.dart';
 import 'package:path/path.dart' as path;
 import 'package:path_provider/path_provider.dart';
 import 'package:pro_image_editor/pro_image_editor.dart';
@@ -882,43 +883,10 @@ class VideoEditorRenderService {
     );
 
     final customTracks = parameters?.audioTracks ?? const <AudioTrack>[];
-
-    final audioTracks = <VideoAudioTrack>[];
-    for (final track in customTracks) {
-      try {
-        final audioPath = await track.audio.safeFilePath();
-        audioTracks.add(
-          VideoAudioTrack(
-            path: audioPath,
-            startTime: track.startTime,
-            endTime: track.endTime,
-            audioStartTime: track.audioStartTime,
-            audioEndTime: track.audioEndTime,
-            loop: track.loop,
-            volume: track.volume,
-          ),
-        );
-      } catch (e, stackTrace) {
-        // Skip an unresolvable track (e.g. failed network download) instead of
-        // aborting the whole render — but log it so the gap is visible.
-        Log.error(
-          'Failed to resolve audio track ${track.id} for render — skipping it',
-          name: _logName,
-          category: .video,
-          error: e,
-          stackTrace: stackTrace,
-        );
-      }
-    }
-
-    if (customTracks.isNotEmpty && audioTracks.isEmpty) {
-      Log.warning(
-        'Render produced no usable audio from ${customTracks.length} '
-        'requested track(s); custom audio will be missing from the output',
-        name: _logName,
-        category: .video,
-      );
-    }
+    final audioTracks = await resolveRenderAudioTracks(
+      customTracks,
+      logName: _logName,
+    );
 
     final volumeSegments = segments
         .map((s) => s.copyWith(volume: s.volume))
