+++ mobile/lib/widgets/video_editor/timeline_editor/video_editor_timeline.dart
@@ -13,13 +13,9 @@ import 'package:openvine/models/timeline_overlay_item.dart';
 import 'package:openvine/widgets/video_editor/main_editor/video_editor_scope.dart';
 import 'package:openvine/widgets/video_editor/timeline_editor/controls/video_editor_timeline_control_bar.dart';
 import 'package:openvine/widgets/video_editor/timeline_editor/strips/video_editor_timeline_clip_strip.dart';
-import 'package:openvine/widgets/video_editor/timeline_editor/strips/video_editor_timeline_overlay_strip.dart';
-import 'package:openvine/widgets/video_editor/timeline_editor/video_editor_timeline_body.dart';
 import 'package:openvine/widgets/video_editor/timeline_editor/video_editor_timeline_geometry.dart';
 import 'package:openvine/widgets/video_editor/timeline_editor/video_editor_timeline_header.dart';
-import 'package:openvine/widgets/video_editor/timeline_editor/video_editor_timeline_markers.dart';
-import 'package:openvine/widgets/video_editor/timeline_editor/video_editor_timeline_playhead.dart';
-import 'package:openvine/widgets/video_editor/timeline_editor/video_editor_timeline_volume.dart';
+import 'package:openvine/widgets/video_editor/timeline_editor/video_editor_timeline_interactive_body.dart';
 
 /// Interactive timeline editor for composing video clips.
 ///
@@ -257,7 +253,7 @@ class _VideoEditorTimelineState extends State<VideoEditorTimelineScaffold> {
                     alignment: .topCenter,
                     minHeight: TimelineConstants.height,
                     maxHeight: TimelineConstants.height,
-                    child: _TimelineInteractiveBody(
+                    child: VideoEditorTimelineInteractiveBody(
                       playheadPosition: _playheadPosition,
                       totalDuration: totalDuration,
                       formatPosition: (pos) {
@@ -911,264 +907,5 @@ class _VideoEditorTimelineState extends State<VideoEditorTimelineScaffold> {
 
     final position = _scrollOffsetToPosition(_scrollController.offset);
     context.read<VideoEditorMainBloc>().add(VideoEditorSeekRequested(position));
-  }
-}
-
-class _TimelineInteractiveBody extends StatelessWidget {
-  const _TimelineInteractiveBody({
-    required this.playheadPosition,
-    required this.totalDuration,
-    required this.formatPosition,
-    required this.onStepPosition,
-    required this.onPointerDown,
-    required this.onPointerMove,
-    required this.onPointerUp,
-    required this.onPointerCancel,
-    required this.onScrollNotification,
-    required this.scrollController,
-    required this.isPinching,
-    required this.isTrimming,
-    required this.halfScreen,
-    required this.pixelsPerSecond,
-    required this.clips,
-    required this.totalWidth,
-    required this.isInteracting,
-    required this.onReorder,
-    required this.onReorderChanged,
-    required this.trimmingClipId,
-    required this.onTrimChanged,
-    required this.onTrimDragChanged,
-    required this.onClipTapped,
-    required this.isMultiSelectMode,
-    required this.selectedClipIds,
-    required this.onOverlayItemMoved,
-    required this.onOverlayItemMoving,
-    required this.onOverlayItemTrimmed,
-    required this.onOverlayTrimDragChanged,
-    required this.onOverlayItemTapped,
-    required this.onOverlayDragStarted,
-    required this.onOverlayDragEnded,
-    required this.verticalScrollController,
-    required this.overlayStripsScrollController,
-    required this.volumePreviewNotifier,
-  });
-
-  final ValueNotifier<Duration> playheadPosition;
-  final Duration totalDuration;
-  final String Function(Duration) formatPosition;
-  final void Function(Duration, Duration, Duration) onStepPosition;
-  final PointerDownEventListener onPointerDown;
-  final PointerMoveEventListener onPointerMove;
-  final PointerUpEventListener onPointerUp;
-  final PointerCancelEventListener onPointerCancel;
-  final bool Function(ScrollNotification) onScrollNotification;
-  final ScrollController scrollController;
-  final bool isPinching;
-  final bool isTrimming;
-  final double halfScreen;
-  final double pixelsPerSecond;
-  final List<DivineVideoClip> clips;
-  final double totalWidth;
-  final bool isInteracting;
-  final void Function(List<DivineVideoClip>) onReorder;
-  final ValueChanged<bool> onReorderChanged;
-  final String? trimmingClipId;
-  final ClipTrimCallback onTrimChanged;
-  final ValueChanged<bool> onTrimDragChanged;
-  final ValueChanged<int> onClipTapped;
-  final bool isMultiSelectMode;
-  final Set<String> selectedClipIds;
-  final OverlayMoveCallback onOverlayItemMoved;
-  final OverlayMovingCallback onOverlayItemMoving;
-  final OverlayTrimCallback onOverlayItemTrimmed;
-  final ValueChanged<bool> onOverlayTrimDragChanged;
-  final ValueChanged<TimelineOverlayItem> onOverlayItemTapped;
-  final ValueChanged<TimelineOverlayItem> onOverlayDragStarted;
-  final VoidCallback onOverlayDragEnded;
-  final ScrollController verticalScrollController;
-  final ScrollController overlayStripsScrollController;
-  final ValueNotifier<double?> volumePreviewNotifier;
-
-  static const _scrollBottomPadding = 100;
-
-  @override
-  Widget build(BuildContext context) {
-    final isVolumeEditMode = context.select(
-      (VideoEditorMainBloc b) => b.state.isVolumeEditMode,
-    );
-    final soundItemCount = context.select(
-      (TimelineOverlayBloc b) => b.state.items
-          .where((i) => i.type == TimelineOverlayType.sound)
-          .length,
-    );
-    final rawVolumeContentHeight =
-        TimelineConstants.rulerHeight +
-        4 +
-        clips.length *
-            (TimelineConstants.thumbnailStripHeight +
-                TimelineConstants.thumbnailVerticalRowGap) -
-        TimelineConstants.thumbnailVerticalRowGap +
-        4 +
-        TimelineConstants.overlayStripGap +
-        soundItemCount * TimelineConstants.soundOverlayRowHeight;
-    final volumeContentHeight =
-        rawVolumeContentHeight > TimelineConstants.height
-        ? rawVolumeContentHeight
-        : TimelineConstants.height;
-
-    return Stack(
-      fit: .expand,
-      children: [
-        SingleChildScrollView(
-          controller: verticalScrollController,
-          physics: isVolumeEditMode
-              ? const ClampingScrollPhysics()
-              : const NeverScrollableScrollPhysics(),
-          padding: .only(
-            bottom: _scrollBottomPadding + MediaQuery.paddingOf(context).bottom,
-          ),
-          child: ClipRect(
-            child: AnimatedContainer(
-              duration: const Duration(milliseconds: 250),
-              curve: Curves.easeInOut,
-              height: isVolumeEditMode
-                  ? volumeContentHeight
-                  : TimelineConstants.height,
-              child: Stack(
-                clipBehavior: Clip.none,
-                children: [
-                  ValueListenableBuilder<Duration>(
-                    valueListenable: playheadPosition,
-                    builder: (context, position, child) {
-                      final increased = Duration(
-                        milliseconds: (position + const Duration(seconds: 1))
-                            .inMilliseconds
-                            .clamp(0, totalDuration.inMilliseconds),
-                      );
-                      final decreased = Duration(
-                        milliseconds: (position - const Duration(seconds: 1))
-                            .inMilliseconds
-                            .clamp(0, totalDuration.inMilliseconds),
-                      );
-                      return Semantics(
-                        label:
-                            context.l10n.videoEditorVideoTimelineSemanticLabel,
-                        slider: true,
-                        value: formatPosition(position),
-                        increasedValue: formatPosition(increased),
-                        decreasedValue: formatPosition(decreased),
-                        onIncrease: () => onStepPosition(
-                          position,
-                          totalDuration,
-                          const Duration(seconds: 1),
-                        ),
-                        onDecrease: () => onStepPosition(
-                          position,
-                          totalDuration,
-                          const Duration(seconds: -1),
-                        ),
-                        child: child ?? const SizedBox.shrink(),
-                      );
-                    },
-                    child: Listener(
-                      onPointerDown: onPointerDown,
-                      onPointerMove: onPointerMove,
-                      onPointerUp: onPointerUp,
-                      onPointerCancel: onPointerCancel,
-                      child: NotificationListener<ScrollNotification>(
-                        onNotification: onScrollNotification,
-                        child: SingleChildScrollView(
-                          controller: scrollController,
-                          scrollDirection: Axis.horizontal,
-                          physics: isPinching || isTrimming
-                              ? const NeverScrollableScrollPhysics()
-                              : const ClampingScrollPhysics(),
-                          clipBehavior: .none,
-                          padding: .symmetric(horizontal: halfScreen),
-                          child: VideoEditorTimelineBody(
-                            totalDuration: totalDuration,
-                            pixelsPerSecond: pixelsPerSecond,
-                            scrollController: scrollController,
-                            overlayStripsScrollController:
-                                overlayStripsScrollController,
-                            scrollPadding: halfScreen,
-                            clips: clips,
-                            totalWidth: totalWidth,
-                            isInteracting: isInteracting,
-                            onReorder: onReorder,
-                            onReorderChanged: onReorderChanged,
-                            trimmingClipId: trimmingClipId,
-                            onTrimChanged: onTrimChanged,
-                            onTrimDragChanged: onTrimDragChanged,
-                            onClipTapped: isVolumeEditMode
-                                ? null
-                                : onClipTapped,
-                            isMultiSelectMode: isMultiSelectMode,
-                            selectedClipIds: selectedClipIds,
-                            onOverlayItemMoved: onOverlayItemMoved,
-                            onOverlayItemMoving: onOverlayItemMoving,
-                            onOverlayItemTrimmed: onOverlayItemTrimmed,
-                            onOverlayTrimDragChanged: onOverlayTrimDragChanged,
-                            onOverlayItemTapped: isVolumeEditMode
-                                ? null
-                                : onOverlayItemTapped,
-                            onOverlayDragStarted: onOverlayDragStarted,
-                            onOverlayDragEnded: onOverlayDragEnded,
-                            playheadPosition: playheadPosition,
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                  Positioned(
-                    left: 0,
-                    top: 0,
-                    child: ClipRect(
-                      child: AnimatedSwitcher(
-                        duration: const Duration(milliseconds: 220),
-                        transitionBuilder: (child, animation) =>
-                            SlideTransition(
-                              position:
-                                  Tween<Offset>(
-                                    begin: const Offset(-1.0, 0.0),
-                                    end: Offset.zero,
-                                  ).animate(
-                                    CurvedAnimation(
-                                      parent: animation,
-                                      curve: Curves.easeInOut,
-                                    ),
-                                  ),
-                              child: child,
-                            ),
-                        layoutBuilder: (currentChild, previousChildren) =>
-                            Stack(
-                              alignment: Alignment.centerLeft,
-                              children: <Widget>[
-                                ...previousChildren,
-                                ?currentChild,
-                              ],
-                            ),
-                        child: isVolumeEditMode
-                            ? VideoEditorTimelineVolume(
-                                key: const ValueKey('volume'),
-                                volumePreviewNotifier: volumePreviewNotifier,
-                              )
-                            : const SizedBox.shrink(key: ValueKey('empty')),
-                      ),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
-        VideoEditorTimelineMarkers(
-          scrollController: scrollController,
-          scrollPadding: halfScreen,
-          pixelsPerSecond: pixelsPerSecond,
-        ),
-        const VideoEditorTimelinePlayhead(),
-      ],
-    );
   }
 }
